Rename code

+++ lib/common/widgets/base_text_widget.dart
@@ -5,7 +5,7 @@ import 'package:yt_ulearning/common/values/colors.dart';
 Widget reusableText(
   String text, {
   Color color = AppColors.primaryText,
-  int fontSize = 16,
+  double fontSize = 16,
   FontWeight fontWeight = FontWeight.bold,
 }) {
   return Text(
